backend/requirements.txt
Django==4.2
djangorestframework==3.15.2
Pillow==10.4.0
pytesseract==0.3.13
celery==5.4.0
redis==5.0.8
psycopg2-binary==2.9.9